+++ src/lib/FormGroupLayout.tsx
@@ -1,21 +1,20 @@
-import { PropsWithChildren } from "react";
+import { PropsWithChildren, ReactNode } from "react";
 import { FieldError, FieldValues, get, useFormContext } from "react-hook-form";
-import { FormGroup, FormFeedback, FormText } from "reactstrap";
+import { FormGroup, FormFeedback, FormText, InputGroup } from "reactstrap";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { CommonInputProps } from "./types/CommonInputProps";
 import "./styles/FormGroupLayout.css";
 import { FormGroupLayoutLabel } from "./FormGroupLayoutLabel";
-import { IconDefinition } from "@fortawesome/fontawesome-svg-core";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 interface FormGroupLayoutProps<T extends FieldValues>
   extends PropsWithChildren<Pick<CommonInputProps<T>, "helpText" | "label" | "name" | "id" | "labelToolTip" | "inputOnly">> {
   layout?: "checkbox" | "switch";
-  icon?: IconDefinition;
+  addonLeft?: ReactNode;
+  addonRight?: ReactNode;
 }
 
 const FormGroupLayout = <T extends FieldValues>(props: FormGroupLayoutProps<T>) => {
-  const { label, helpText, children, layout, labelToolTip, inputOnly, icon } = props;
+  const { label, helpText, children, layout, labelToolTip, inputOnly, addonLeft, addonRight } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const {
     formState: { errors },
@@ -39,21 +38,11 @@ const FormGroupLayout = <T extends FieldValues>(props: FormGroupLayoutProps<T>) 
   ) : (
     <FormGroup switch={switchLayout ? true : undefined} check={checkboxLayout ? true : undefined}>
       <FormGroupLayoutLabel<T> label={label} fieldName={name} fieldId={id} tooltip={labelToolTip} layout={layout} />
-      {children}
-      {icon && (
-        <FontAwesomeIcon
-          icon={icon}
-          fixedWidth
-          size="lg"
-          style={{
-            float: "right",
-            marginRight: "10px",
-            marginTop: "-27px",
-            position: "relative",
-            zIndex: "2",
-          }}
-        />
-      )}
+      <InputGroup>
+        {addonLeft}
+        {children}
+        {addonRight}
+      </InputGroup>
       <FormFeedback>{errorMessage}</FormFeedback>
       {helpText && <FormText>{helpText}</FormText>}
     </FormGroup>