+++ src/lib/FormattedInput.tsx
@@ -26,7 +26,8 @@ const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) =>
     onBlur: propsOnBlur,
     labelToolTip,
     markAllOnFocus,
-    icon,
+    addonLeft,
+    addonRight,
   } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const { control } = useFormContext();
@@ -52,7 +53,15 @@ const FormattedInput = <T extends FieldValues>(props: FormattedInputProps<T>) =>
         };
 
         return (
-          <FormGroupLayout helpText={helpText} name={name} id={id} label={label} labelToolTip={labelToolTip} icon={icon}>
+          <FormGroupLayout
+            helpText={helpText}
+            name={name}
+            id={id}
+            label={label}
+            labelToolTip={labelToolTip}
+            addonLeft={addonLeft}
+            addonRight={addonRight}
+          >
             <>
               {numericFormat && (
                 <NumericFormat